strategy_btc_price_prediction/requirements.txt
py-clob-client
python-dotenv
pandas
numpy
requests
